+++ ext/monitor/src/main/java/app/MonitorApp.java
@@ -82,7 +82,10 @@ public class MonitorApp extends App {
     }
 
     private void configureSlackClient() {
-        bind(HTTPClient.class, new HTTPClientBuilder().build());
+        bind(HTTPClient.class, new HTTPClientBuilder()
+                .maxRetries(3)
+                .retryWaitTime(Duration.ofSeconds(2))   // slack has rate limit with 1 message per second, here to slow down further when hit limit, refer to https://api.slack.com/docs/rate-limits
+                .build());
 
         Bean.register(SlackMessageAPIRequest.class);
         Bean.register(SlackMessageAPIResponse.class);