+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Change log
 ### 7.3.5 (1/30/2020 - )
-* 
+* httpClient: support retry on 429 too many requests
 
 ### 7.3.4 (1/29/2020 - 1/30/2020)
 * es: update to 7.5.2